refactor(ui): TMS-2508 rename simpledialog
Renamed DialogBox to SimpleDialog.
