+++ frontend/src/components/ui/SimpleDialog.vue
